+++ ui/desktop/electron-app/src/index.js
@@ -21,6 +21,7 @@ const sessionManager = require('./services/session-manager.js');
 
 const menu = require('./config/menu.js');
 const appUpdater = require('./helpers/app-updater.js');
+const { isWindows } = require('./helpers/platform.js');
 const isDev = require('electron-is-dev');
 
 // Register the custom file protocol
@@ -115,6 +116,7 @@ app.on('ready', async () => {
     width: 1280,
     height: 760,
     titleBarStyle: 'hiddenInset',
+    frame: !isWindows(),
     webPreferences: {
       partition,
       sandbox: true,
@@ -195,16 +197,8 @@ app.on('before-quit', (event) => {
       detail: 'Close sessions before quitting?',
     };
 
-    dialog.showMessageBox(dialogOpts).then((returnValue) => {
-      switch (returnValue.response) {
-        case 0:
-          sessionManager.stopAll();
-          break;
-        default:
-          event.preventDefault();
-          break;
-      }
-    });
+    const buttonId = dialog.showMessageBoxSync(null, dialogOpts);
+    buttonId === 0 ? sessionManager.stopAll() : event.preventDefault();
   }
 });