+++ ui/desktop/electron-app/src/models/session.js
@@ -1,5 +1,7 @@
-const { spawnAsyncJSONPromise } = require('../helpers/spawn-promise.js');
+const treeKill = require('tree-kill');
 const sanitizer = require('../utils/sanitizer.js');
+const { isWindows } = require('../helpers/platform.js');
+const { spawnAsyncJSONPromise } = require('../helpers/spawn-promise.js');
 
 class Session {
   #id;
@@ -63,7 +65,14 @@ class Session {
       if (this.isRunning) {
         this.#process.on('close', () => resolve());
         this.#process.on('error', (e) => reject(e));
-        this.#process.kill();
+        /**
+         * On Windows OS, a spawned process uses cmd.exe to initiate a session.
+         * Hence, captured process.pid corresponds to cmd.exe instead of session.
+         * To avoid orphaned session processes and due to lack of node support
+         * to handle killing processes cleanly in this scenario,
+         * kill entire dependent process tree on Windows.
+         */
+        isWindows() ? treeKill(this.#process.pid) : this.#process.kill();
       } else {
         // Do nothing when process isn't running
         resolve();