+++ ui/desktop/tests/helpers/window-mock-ipc.js
@@ -17,6 +17,8 @@ class MockIPC {
   }
 
   resetOrigin() {}
+  isMacOS() {}
+  isWindowsOS() {}
 }
 
 /**